Updated menu code to paginate. Created a copy of the menu draw for the title listings, because I wanted to pass the results vector to the draw routine rather than duping the data into a screen formatted char. A todo would be to re-unify the code (I've updated the routine on both) and use a call back for the purposes of formatting the output. Moved menu.c to menu.cpp purely for this, and it was better than a separate cpp file just for this.
